requirements.txt
alembic==1.17.0
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.7.14
cffi==2.0.0
charset-normalizer==3.4.2
click==8.3.0
colorama==0.4.6
contourpy==1.3.2
cryptography==46.0.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
dnspython==2.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.117.1
fonttools==4.59.0
greenlet==3.2.3
h11==0.16.0
httptools==0.6.4
idna==3.10
kiwisolver==1.4.8
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.3
numpy==2.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pyparsing==3.2.3
pypdf==5.0.0
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
psycopg2-binary
pyjwt 
passlib[bcrypt]

